reduce size of large file; ./test passes
